requirements.txt
chardet==4.0.0
cycler==0.10.0
fpdf==1.7.2
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
Pillow==8.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
pyyaml==5.3.1
six==1.15.0
